pyproject.toml
[project]
name = "gator-tune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "async-timeout==5.0.1",
    "attrs==24.2.0",
    "audioop-lts==0.2.2",
    "beautifulsoup4==4.12.3",
    "bs4==0.0.2",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.4.0",
    "colorama==0.4.6",
    "decorator==4.4.2",
    "discord==2.3.2",
    "discord-py==2.6.0",
    "frozenlist==1.5.0",
    "html5lib==1.1",
    "idna==3.10",
    "imageio==2.36.0",
    "imageio-ffmpeg==0.5.1",
    "iniconfig==2.0.0",
    "moviepy==1.0.3",
    "multidict==6.1.0",
    "numpy==2.1.3",
    "packaging==24.1",
    "pillow==11.0.0",
    "pluggy==1.5.0",
    "proglog==0.1.10",
    "propcache==0.2.0",
    "pycparser==2.22",
    "pynacl==1.5.0",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "pytube==15.0.0",
    "pytubefix==10.3.6",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "setuptools==80.9.0",
    "six==1.16.0",
    "soupsieve==2.6",
    "tqdm==4.66.6",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "webencodings==0.5.1",
    "yarl==1.17.1",
]
